Add Simperium provider for notes
Before running make sure to replace the APP_ID and API_TOKEN with
values you get from your Simperium.com dashboard!

In this patch we're connecting to a new Simperium app and using a
bucket entity to store the contents of the block editor. This is a
list of block nodes.

Loading of the app takes place after conncting to the bucket and may
be funny on the first load. If that's the case you might need to
manually create the new entity...

```
bucket.update( 'blocks', { blocks: [] } );
```

Afterwards you should be able to connect multiple devices from this
branch in different browsers or even on different machines and the
editor contents should stay consistent.

One issue that makes for a buggy experience is that the Simperium
library is replacing the `blocks` data instead of diffing and merging
it. This means that one edit session carries a risk of wiping out the
changes from another session, and if both are offline when they make
their changes then only one's changes will come though. This should
be fixable but I didn't have time in this patch to resolve it.